refactor: @constants/styles.ts 에서 사용하던 함수를 style/index.ts로 이동
[#360]
